+++ src/config.ts
@@ -1,4 +1,3 @@
-import Cache from '@oclif/core/lib/cache'
 import type { ActionBase } from './action/base'
 import simple from './action/simple'
 import spinner from './action/spinner'
@@ -18,7 +17,7 @@ const actionType =
   (Boolean(process.stderr.isTTY) &&
     !process.env.CI &&
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    !['dumb', 'emacs-color'].includes(process.env.TERM!) &&
+    !['dumb', 'emacs-color'].includes(process.env.TERM || '') &&
     'spinner') ||
   'simple'
 
@@ -52,11 +51,16 @@ export class Config {
 }
 
 function fetch(): any {
+  /*
   const core = Cache.getInstance().get('@oclif/core')
   const major = core?.version.split('.')[0] || 'unknown'
   if (globals[major]) return globals[major]
   globals[major] = new Config()
   return globals[major]
+*/
+  if (globals.config) return globals.config
+  globals.config = new Config()
+  return globals.config
 }